https://www.rfc-editor.org/errata/eid5775
TODO: We need to decide what to do here.

Perhaps "0" is a good value for "no compound MAC is used"?

This field should also be described as a sequence of bits, instead
of as a 4-bit integer.  i.e. "E | M | R | R"

E == EMSK compond MAC is present
M == MSK compound MAC is present
R == reserved

The "Length" field should also not be a fixed size.  If one (or both)
of the compound MACs is missing, then the TLV can perhaps omit them?

Also, what happens if the MACs are not 20 octets, as described in

	https://www.rfc-editor.org/errata/eid5768

what happens then?