Change concat to function param
The TLS PRF is defined as 
PRF(secret, label, seed) = P_<hash>(secret, label | seed)
If we reference the PRF it should use the seed as a parameter instead of concatenation.  The TLS-PRF is described earlier in the section, but we can duplicate it here if it makes it clearer.